Moved array functions
