+++ src/functions/array/arraySort.js
@@ -1,0 +1,16 @@
+module.exports = async (d) => {
+    const data = d.util.aoiFunc(d);
+    if (data.err) return d.error(data.err);
+    const [name, type = "asc"] = data.inside.splits;
+    if (!d.data.arrays[name])
+        return d.aoiError.fnError(d, "custom", { inside: data.inside }, `Array with name "${name}" does not exists.`);
+    d.arrays[name] = d.arrays[name].sort((a, b) =>
+        type === "asc" ? a - b : b - a,
+    );
+    d.data.arrays = d.arrays;
+    return {
+        code: d.util.setCode(data),
+        arrays: d.arrays,
+        data: d.data,
+    };
+};
